Replace C-style casts
C-style casts are error prone and hard to find. This removes
all unnecessary casts and replaces the remaining C-style
casts with C++-style casts.

+++ hdtv/rootext/display/DisplaySpec.hh
@@ -86,8 +86,8 @@ public:
   bool GetDrawUnderflowBin() { return fDrawUnderflowBin; }
   bool GetDrawOverflowBin() { return fDrawOverflowBin; }
 
-  double GetMinCh(void) { return (double)fHist->GetXaxis()->GetXmin(); }
-  double GetMaxCh(void) { return (double)fHist->GetXaxis()->GetXmax(); }
+  double GetMinCh(void) { return fHist->GetXaxis()->GetXmin(); }
+  double GetMaxCh(void) { return fHist->GetXaxis()->GetXmax(); }
 
   Int_t ClipBin(Int_t bin) {
     if (fDrawUnderflowBin) {
